% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_increment.R
\name{increment_month}
\alias{increment_month}
\title{This function increments a date to the first
or last date of the next month.}
\usage{
increment_month(date_to_increment, start_or_end)
}
\arguments{
\item{date_to_increment}{The date to work on.}

\item{start_or_end}{Return the date of the start or the end of the next month.}
}
\value{
The updated date.
}
\description{
This function increments a date to the first
or last date of the next month.
}
